dpll_cascading: remove unused CM_SHADOW_FREQ_CONFIG2
The value from CM_SHADOW_FREQ_CONFIG2 can't be applied because
GPMC is always active and can't be entered in idle state.

The GPMC is always active because on OMAP4, CPU accesses on
unmapped addresses are redirected to GPMC by L3 interconnect.
CPU is speculative nature and such accesses are possible which
can lead to indirect access to GPMC and if it's clock is
not running, it can result in hang/abort on the platform.
See path https://github.com/Ntemis/android_kernel_bn_omap/commit/d60565948468b for more information.