regulator: Correct voltage value used in regulator_set_voltage_complete
The voltage value specified in the tracepoint function
trace_regulator_set_voltage_complete() inside of the function
_regulator_do_set_voltage() is erroneously set to -1 if a
list_voltage() callback is not specified by a regulator device
even if a get_voltage() callback is present.

Add a check for the get_voltage() callback function and call it
in order to determine the voltage that was chosen from the min_uV
to max_uV range.