mm, oom: fix badness score underflow
If the privileges given to root threads (3% of allowable memory) or a
negative value of /proc/pid/oom_score_adj happen to exceed the amount of
rss of a thread, its badness score overflows as a result of commit
a7f638f999ff ('mm, oom: normalize oom scores to oom_score_adj scale only for userspace').

Fix this by making the type signed and return 1, meaning the thread is
still eligible for kill, if the value is negative.