It seems that there is no external PCB temperature sensor is included
This means the thermal pcb domain is not reached on 4430 devices and causes useless
warnings which fill up the logcat.
On 4460 devices it causes a wrong calculated temperature, which means the cpu freq driver
isn't able to react with throttling down the CPU.

With this patch the thermal_lookup request is silently aborted with an ENODEV error, which
is handled at the further thermal calculation routines.